pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp_duckduckgo"
version = "0.1.0"
description = "DuckDuckGo search plugin for Model Context Protocol"
authors = [
    {name = "Gianluca Mazza", email = "info@gianlucamazza.it"}
]
readme = "README.md"
requires-python = ">=3.9"
license = {file = "LICENSE"}
keywords = ["mcp", "duckduckgo", "search", "claude", "anthropic", "llm", "ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.23.2",
    "pydantic>=2.4.2",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "beautifulsoup4>=4.12.0",
    "mcp>=1.3.0",
]

[project.urls]
Homepage = "https://github.com/gianlucamazza/mcp_duckduckgo"
Repository = "https://github.com/gianlucamazza/mcp_duckduckgo"
"Bug Tracker" = "https://github.com/gianlucamazza/mcp_duckduckgo/issues"
Documentation = "https://github.com/gianlucamazza/mcp_duckduckgo#readme"

[project.scripts]
mcp-duckduckgo = "mcp_duckduckgo.main:main"

[tool.black]
line-length = 88
target-version = ["py39"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
